working on budget

+++ db/migrate/20220104022530_create_budget_items.rb
@@ -3,8 +3,8 @@ class CreateBudgetItems < ActiveRecord::Migration[6.1]
     create_table :budget_items do |t|
       t.belongs_to :user, null: false, foreign_key: true
       t.string :category
-      t.integer :total_budget
-      t.integer :current_total
+      t.float :total_budget
+      t.float :current_total
 
       t.timestamps
     end
